requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
"numpy<2"
pyarrow>=14.0.0
plotly>=5.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
